web/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openpyxl==3.1.2
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-ldap==3.4.3
requests==2.31.0
urllib3==2.0.3
Werkzeug==2.3.6
WTForms==3.0.1
xlrd==2.0.1